+++ lib/expandable_fab/expandable_fab.dart
@@ -1,30 +1,63 @@
 import 'package:animated_expandable_fab/animated_expandable_fab.dart';
+import 'package:animated_expandable_fab/expandable_fab/tap_to_open_fab.dart';
 import 'package:flutter/material.dart';
 
+import 'tap_to_close_fab.dart';
 import 'expanding_action_button.dart';
 
-// ExpandableFab is a stateful widget that allows creating an expandable Floating Action Button
-// which can contain multiple action buttons inside it.
+/// A stateful widget that creates an expandable Floating Action Button (FAB).
+///
+/// This widget allows the integration of multiple action buttons within a single FAB.
+/// The FAB can toggle between an open and closed state, revealing or hiding the action buttons.
 class ExpandableFab extends StatefulWidget {
-  const ExpandableFab({
-    super.key,
-    // boolean variable to determine if the fab should be open or closed by default
-    this.initialOpen = false,
-    // the distance between the fab and the action buttons
-    required this.distance,
-    // the list of action buttons to be added to the fab
-    required this.children,
-    // the icon to be displayed when the fab is closed
-    this.closeIcon,
-    // the icon to be displayed when the fab is open
-    this.openIcon,
-  });
-
+  /// Determines whether the FAB is open or closed initially.
   final bool initialOpen;
+
+  /// The distance between the main FAB and the expanded action buttons.
   final double distance;
+
+  /// A list of action buttons that are displayed when the FAB is expanded.
   final List<ActionButton> children;
+
+  /// The icon displayed on the FAB when it is in its closed state.
+  /// If null, a default icon may be used.
   final Widget? closeIcon;
+
+  /// The icon displayed on the FAB when it is in its open state.
+  /// If null, a default icon may be used.
   final Widget? openIcon;
+
+  /// The elevation of the FAB when it is in the open state.
+  /// If null, a default elevation may be applied.
+  final double? openElevation;
+
+  /// The elevation of the FAB when it is in the closed state.
+  /// Defaults to 4.0 if not specified.
+  final double closeElevation;
+
+  /// The color of the shadow cast by the FAB when it is in the closed state.
+  /// If null, the default shadow color is used.
+  final Color? closeShadowColor;
+
+  /// The background color of the FAB when it is in the closed state.
+  /// If null, a default color is used.
+  final Color? closeBackgroundColor;
+
+  /// Constructs an [ExpandableFab].
+  ///
+  /// Requires [distance] and [children]. Other parameters are optional and have default values.
+  const ExpandableFab({
+    this.closeBackgroundColor,
+    this.closeElevation = 4.0,
+    this.initialOpen = false,
+    required this.children,
+    required this.distance,
+    this.closeShadowColor,
+    this.openElevation,
+    this.closeIcon,
+    this.openIcon,
+    super.key,
+  });
 
   @override
   State<ExpandableFab> createState() => _ExpandableFabState();
@@ -34,16 +67,14 @@ class _ExpandableFabState extends State<ExpandableFab>
     with SingleTickerProviderStateMixin {
   late final AnimationController _controller;
   late final Animation<double> _expandAnimation;
-  bool _open = false;
+  late bool _isOpen = widget.initialOpen;
 
   @override
   void initState() {
     super.initState();
-    // initializing the state of the fab
-    _open = widget.initialOpen;
     _controller = AnimationController(
       // setting the initial value of the animation
-      value: _open ? 1.0 : 0.0,
+      value: _isOpen ? 1.0 : 0.0,
       duration: const Duration(milliseconds: 250),
       vsync: this,
     );
@@ -54,25 +85,6 @@ class _ExpandableFabState extends State<ExpandableFab>
     );
   }
 
-  // dispose the animation controller when the widget is disposed
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  // toggle the state of the fab when it is tapped
-  void _toggle() {
-    setState(() {
-      _open = !_open;
-      if (_open) {
-        _controller.forward();
-      } else {
-        _controller.reverse();
-      }
-    });
-  }
-
   // the main build function
   @override
   Widget build(BuildContext context) {
@@ -81,43 +93,40 @@ class _ExpandableFabState extends State<ExpandableFab>
         alignment: Alignment.bottomRight,
         clipBehavior: Clip.none,
         children: [
-          _buildTapToCloseFab(),
+          TapToCloseFab(
+            closeBackgroundColor: widget.closeBackgroundColor,
+            closeShadowColor: widget.closeShadowColor,
+            closeElevation: widget.closeElevation,
+            closeIcon: closeIcon,
+            open: _isOpen,
+            // Optional
+            toggle: _toggleFab,
+          ),
           ..._buildExpandingActionButtons(),
-          _buildTapToOpenFab(),
+          TapToOpenFab(
+            openIcon: widget.openIcon,
+            openElevation: widget.openElevation,
+            toggle: _toggleFab,
+            open: _isOpen,
+          )
         ],
       ),
     );
   }
 
-  // the tap-to-close fab widget
-  Widget _buildTapToCloseFab() {
-    return SizedBox(
-      width: 56.0,
-      height: 56.0,
-      child: Center(
-        child: Material(
-          shape: const CircleBorder(),
-          clipBehavior: Clip.antiAlias,
-          elevation: 4.0,
-          child: InkWell(
-            onTap: _toggle,
-            child: Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: _open == false
-                  ? const SizedBox()
-                  : (widget.closeIcon ??
-                      Icon(
-                        Icons.close,
-                        color: Theme.of(context).primaryColor,
-                      )),
-            ),
-          ),
-        ),
-      ),
-    );
+  // toggle the state of the fab when it is tapped
+  void _toggleFab() {
+    setState(() {
+      _isOpen = !_isOpen;
+      if (_isOpen) {
+        _controller.forward();
+      } else {
+        _controller.reverse();
+      }
+    });
   }
 
-// builds the expanding action buttons
+  // builds the expanding action buttons
   List<Widget> _buildExpandingActionButtons() {
     final children = <Widget>[];
     final count = widget.children.length;
@@ -133,7 +142,7 @@ class _ExpandableFabState extends State<ExpandableFab>
           child: widget.children[i],
           onTap: () {
             if (widget.children[i].closeFabOnTap) {
-              _toggle();
+              _toggleFab();
             }
             widget.children[i].onPressed?.call();
           },
@@ -143,30 +152,18 @@ class _ExpandableFabState extends State<ExpandableFab>
     return children;
   }
 
-  Widget _buildTapToOpenFab() {
-    return IgnorePointer(
-      ignoring: _open,
-      child: AnimatedContainer(
-        transformAlignment: Alignment.center,
-        transform: Matrix4.diagonal3Values(
-          _open ? 0.7 : 1.0,
-          _open ? 0.7 : 1.0,
-          1.0,
-        ),
-        duration: const Duration(milliseconds: 250),
-        curve: const Interval(0.0, 0.5, curve: Curves.easeOut),
-        child: AnimatedOpacity(
-          opacity: _open ? 0.0 : 1.0,
-          curve: const Interval(0.25, 1.0, curve: Curves.easeInOut),
-          duration: const Duration(milliseconds: 250),
-          child: FloatingActionButton(
-            onPressed: _toggle,
-            child: _open
-                ? const SizedBox()
-                : (widget.openIcon ?? const Icon(Icons.add)),
-          ),
-        ),
-      ),
-    );
+  Widget get closeIcon => _isOpen == false
+      ? const SizedBox()
+      : (widget.closeIcon ??
+          Icon(
+            Icons.close,
+            color: Theme.of(context).primaryColor,
+          ));
+
+  // dispose the animation controller when the widget is disposed
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
   }
 }