+++ CHANGELOG.md
@@ -5,3 +5,10 @@ Includes the ExpandableFab widget
 Allows for customization of open and close icon, children to be displayed when expanded, and initial open state
 Provides functionality for expanding and closing the floating action button on tap of the open/close icon or children
 Includes example usage and tests in the package
+
+
+### 0.0.3
+
+**Enhancements:**
+- Added support for displaying a customizable text widget in addition to the icon for `ActionButton`.
+- Enhanced customization options for buttons' colors and elevations.